Refactoring: Introduced Edge License service. Remove web dependency from dao
